+++ tket/src/Gate/Gate.cpp
@@ -110,6 +110,8 @@ Op_ptr Gate::dagger() const {
     case OpType::Rx:
     case OpType::PhaseGadget:
     case OpType::CnRy:
+    case OpType::CnRx:
+    case OpType::CnRz:
     case OpType::XXPhase:
     case OpType::YYPhase:
     case OpType::ZZPhase:
@@ -117,6 +119,15 @@ Op_ptr Gate::dagger() const {
     case OpType::ISWAP:
     case OpType::ESWAP: {
       return get_op_ptr(optype, minus_times(params_[0]), n_qubits_);
+    }
+    case OpType::GPI: {
+      return get_op_ptr(optype, params_[0]);
+    }
+    case OpType::GPI2: {
+      return get_op_ptr(optype, params_[0] + 1);
+    }
+    case OpType::AAMS: {
+      return get_op_ptr(optype, {params_[0], params_[1] + 1, params_[2]});
     }
     case OpType::ZZMax: {
       // ZZMax.dagger = ZZPhase(-0.5)
@@ -206,6 +217,8 @@ Op_ptr Gate::transpose() const {
     case OpType::SXdg:
     case OpType::CRz:
     case OpType::CRx:
+    case OpType::CnRz:
+    case OpType::CnRx:
     case OpType::CU1:
     case OpType::U1:
     case OpType::Rz:
@@ -231,6 +244,13 @@ Op_ptr Gate::transpose() const {
     case OpType::CnX:
     case OpType::CnZ: {
       return get_op_ptr(optype, std::vector<Expr>(), n_qubits_);
+    }
+    case OpType::GPI:
+    case OpType::GPI2: {
+      return get_op_ptr(optype, -params_[0]);
+    }
+    case OpType::AAMS: {
+      return get_op_ptr(optype, {params_[0], -params_[1], -params_[2]});
     }
     case OpType::U2: {
       // U2(a,b).transpose() == U2(b+1,a+1)
@@ -367,7 +387,8 @@ std::optional<double> Gate::is_identity() const {
     case OpType::YYPhase:
     case OpType::ZZPhase:
     case OpType::XXPhase3:
-    case OpType::ESWAP: {
+    case OpType::ESWAP:
+    case OpType::AAMS: {
       Expr e = params[0];
       if (equiv_0(e, 4)) {
         return 0.;
@@ -421,7 +442,9 @@ std::optional<double> Gate::is_identity() const {
     case OpType::CRy:
     case OpType::PhaseGadget:
     case OpType::ISWAP:
-    case OpType::CnRy: {
+    case OpType::CnRy:
+    case OpType::CnRx:
+    case OpType::CnRz: {
       return equiv_0(params[0], 4) ? 0. : notid;
     }
     case OpType::FSim: {
@@ -462,6 +485,22 @@ bool Gate::is_clifford() const {
     case OpType::PhasedISWAP:
     case OpType::FSim:
       return equiv_0(4 * params_.at(0)) && equiv_0(2 * params_.at(1));
+    case OpType::GPI:
+      return equiv_0(8 * params_.at(0));
+    case OpType::GPI2:
+      return equiv_0(4 * params_.at(0));
+    case OpType::AAMS:
+      if (equiv_0(params_.at(0))) {
+        return true;
+      } else if (
+          !equiv_0(4 * params_.at(0)) || !equiv_0(8 * params_.at(1)) ||
+          !equiv_0(8 * params_.at(2))) {
+        return false;
+      } else if (equiv_0(2 * params_.at(0))) {
+        return true;
+      } else {
+        return equiv_0(4 * params_.at(1)) && equiv_0(4 * params_.at(2));
+      }
     default:
       return false;
   }
@@ -542,6 +581,8 @@ bool Gate::has_symmetry(unsigned port1, unsigned port2) const {
       // n (+1) qubit gates
     case OpType::CnX:
     case OpType::CnY:
+    case OpType::CnRx:
+    case OpType::CnRz:
     case OpType::CnRy: {
       // symmetry on first n ports not on n+1
       auto last_port = n_q - 1;
@@ -705,6 +746,12 @@ std::vector<Expr> Gate::get_tk1_angles() const {
           params_.at(1) + half, params_.at(0), params_.at(2) - half,
           (params_.at(1) + params_.at(2)) / 2};
     }
+    case OpType::GPI: {
+      return {2 * params_.at(0), 1, 0, half};
+    }
+    case OpType::GPI2: {
+      return {params_.at(0), half, -params_.at(0), 0};
+    }
     case OpType::NPhasedX: {
       if (n_qubits_ != 1) {
         throw BadOpType(
@@ -824,6 +871,7 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
     }
     case OpType::CZ:
     case OpType::CRz:
+    case OpType::CnRz:
     case OpType::CS:
     case OpType::CSdg:
     case OpType::CU1:
@@ -862,6 +910,7 @@ std::optional<Pauli> Gate::commuting_basis(unsigned i) const {
     case OpType::CSX:
     case OpType::CSXdg:
     case OpType::CRx:
+    case OpType::CnRx:
     case OpType::CX:
     case OpType::CCX:
     case OpType::CnX: {